backend/pyproject.toml
[project]
name = "speakingbuddy-backend"
version = "0.1.0"
description = "SpeakingBuddy MVP Backend — Luxembourgish pronunciation learning API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "praat-parselmouth>=0.4.4",
    "pydub>=0.25.1",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
